✨Add alert to slack when Sentry usage over threshold

+++ python/services/sentry_service.py
@@ -7,7 +7,7 @@ from python.clients.sentry_client import SentryClient
 class UsageStats:
     total: int
     max_usage: int
-    percentage_of_quota_used: int
+    percentage_of_quota_used: float
 
 
 class SentryService:
